Test automation change only: fix recursion on clone of a repo. (#452)
The test automation scripts use a utility function, fetch_repo() to pull down full repositories.  This handles recursion _except_, it turns out, in the initial clone step, which is fixed here.  Looks like the functionality was broken some time ago but we didn't notice because the code had previously been cloned on the test agent where it was being run.