Cellular test change only: disconnect from network after uCelLLocLoc. (#474)
SARA-R5 can get into a bit of a tizzy if the module is left connected after the Cell Locate test has done its stuff: when the next test begins the SARA-R5 sometimes just happens to reconnect to the network (visible with a +CSON: 1) in the middle of the initial AT command sequence (AT&Kx etc.) and then becomes unresponsive after the AT+CFUN=4 is sent just a second later at the end of that initialisation sequence. To avoid this corner case, we disconnect the module from the network at the end of the uCellLocLoc test.