Fix to test of port code: apparent strtok_r memory leak. (#468)
The port tests have recently begun failing on nRF5 (test instance 15.0.0) due to an apparent memory leak in strtok_r(). This is a lie; what seems to be happening is that the timing on the nRF5 platform has changed [possibly due to the UART re-write?] and the RTOS idle task is now not getting to run for long enough to clean up the dynamic tasks created in the previous test (portEventQueue()); this can be seen in that the test which follows strtok_r() reports a negative heap loss that is as big as the positive heap loss that is the strtok_r() test failure.

With this commit a delay is added at the end of the portEventQueue() test to give the RTOS idle task time to do its stuff.