Test fix only in AT client: restore stop tag. (#467)
In the AT command test atClientCommandSet1() a slew of AT sequences are sent over back-to-backed UARTs to fully exercise the AT client code. Some of these AT sequences correctly call uAtClientIgnoreStopTag() (when reading parameters that contain binary data that might legitimately include the letters "OK") however the test code failed to note that this had been done and call uAtClientRestoreStopTag() before ending the command; this meant that, depending on timing, it was possible for the "OK" in the AT response to be missed; the "OK" would then be taken as the response to the next AT command to be tried, making it appear as though there was nothing in the response, just an immediate "OK", and causing a test failure.

A subtle shift in timing, resulting from the addition of a test for the up-coming support of critical sections, caused this problem to become visible.

It is now fixed.