Windows platform change and additional port test: uPortTaskBlock() fix. (#471)
The SleepEx() WIN32 API function, used to implement uPortTaskBlock() on Windows, can return early, so this commit adds a check that the right amount of time has elapsed and, if it hasn't, returns to sleep.

This was spotted while watching a test which reboots the module (cellCfgBandMask()); it should have waited 15 seconds for the reboot but actually waited for no time at all. This may explain some of the flakiness we have been seeing in the Windows platform tests, if task blocking periods can sometimes expire immediately.

A test is added in the port tests so that we can check, at least visually in the test log, that the blocking function does work; even if we're not checking accurate timings 'cos it's Windows, it shouldn't be zero.