Cell Locate fix: location not being detected quickly (#470)
In the code that waits for an answer from the Cell Locate server a variable is passed to a callback that has to be checked for an answer.  The variable that is being checked is marked as volatile so that the compiler should not optimise-out the checking but that variable is within a larger structure that is not volatile and it would seem that the whole *structure* needs to be volatile to persuade the compiler to actually do the check, otherwise it waits for ages for no reason.