BREAKING CHANGE IF YOU HAVE YOUR OWN PORT: add critical section and U_INLINE. (#463)
This commit introduces two new port API functions: uPortEnterCritical() and uPortExitCritical(). When uPortEnterCritical() is called all RTOS task [re]scheduling and interrupts are switched off, when uPortExitCritical() is called they are resumed. This will be used within the uAtClientSetWakeUpHandler() feature of the AT client to support the up-coming implementation of cellular power saving.

No OS-related port API calls (i.e. anything to do with tasks, mutexes, semaphores, queues or timer operations) should be made within a critical section; depending on the platform that may cause an assert (FreeRTOS) or may cause the rescheduling that you didn't want to happen anyway (Zephyr).

Since a critical section should be in place for as little time as possible a new macro is introduced in u_compiler.h, U_INLINE, which can be used to force the compiler to in-line a function, so that it is just like a macro, on a given toolchain. U_INLINE is populated for GCC (and hence ARMCC) and resolves to nothing on MSVC (since Windows is powerful enough in any case). U_INLINE is applied to the uPortEnterCritical() and uPortExitCritical() functions.

Customers who have their own port that do not wish to implement uPortEnterCritical() and uPortExitCritical() may have their implementations return U_ERROR_COMMON_NOT_IMPLEMENTED and all will be fine. However, that may mean that certain features of ubxlib (e.g. the up-coming cellular power-saving feature) will NOT operate.