automation: Add workaround for stale JLink locks
If system is rebooted while a JLink lock is currently held it looks like
JLink file locks will not be cleaned up. Since the lock file contains the
PID of the process that took the lock and this can become a problem as
the PID can point at a new process after reboot.