Test/example change only: modify u_runner for ROPI.
When compiling ubxlib as position-independent code (Read Only Position Independent or ROPI) the macro U_PORT_UNITY_TEST_FUNCTION, which collects-up all of the tests and examples, doesn't work, since the address of the constant strings it usually picks up are only resolved by the linker, which isn't actually being used. With this commit the macro is changed (subtly!) so that the compiler ends up populating the values instead.