Test/example change only: remove delay from u_runner test list print.
Some time ago a very short delay, 1 ms, was added between each line printed by the u_runner uRunnerPrintAll() function, so that the RTT logging buffer did not fill up.  With the previous commit to change the way the test name strings are populated by the compiler, this delay actually seems to _cause_ a problem, in that on nRF52840, with the nRF5 toolchain/GCC, RTT logging stops during the printing of the list of tests unless this delay is _removed_, in which case it works fine.

Since the delay was only added to make RTT happy and it turns out to now be making RTT sad, it has been removed.