AT client changes only: uAtClientUrcDirect() and AT+UPSV bug fix. (#478)
* Temporary change only: remove U_CFG_MUTEX_DEBUG.

Introduction of the revised approach for waking up the module from power saving in the AT client has caused a lot more mutex activity and revealed a problem in the implementation of U_CFG_MUTEX_DEBUG.  This is feature is removed from the test farm builds temporarily while the problem is investigated.

* AT client changes only: uAtClientUrcDirect() and AT+UPSV bug fix.

A new API is added to the AT client: uAtClientUrcDirect().  It should NOT normally be necessary to use this function - URCs should normally be handled asynchronously with uAtClientSetUrcHandler().  However, there are cases (e.g. in code handling wake-up from deep sleep) where it is necessary to wait for a URC "in-line".  For those cases, the same URC handler callback as would be written for the uAtClientSetUrcHandler() case, can be called directly using this function, a bit like the usual uAtClientResponseStart()/Stop() mechanism but with no "OK" conclusion and no possibility of an "ERROR" response.

During the development of this feature it was found that the existing UART power saving (AKA AT+UPSV) wake-up mechanism, the code behind uAtClientSetWakeUpHandler(), was not thread-safe: the new, thread-safe, implementation in this commit relies upon the new uPortEnterCritical()/uPortExitCritical() and uPortTaskGetHandle() features of the port API.

A bug was also found in uAtClientSetWakeUpHandler(), where it could end up inserting a spurious comma at the start of an AT command (because a flag was not being saved when entering the wake-up handler and then restored afterwards).  This is now fixed.