Cellular change only: add support for 3GPP power saving and E-DRX. (#479)
Functions are provided to set/get 3GPP sleep and E-DRX (uCellPwrSetRequested3gppPowerSaving(), uCellPwrGetRequested3gppPowerSaving(), uCellPwrGet3gppPowerSaving(), uCellPwrSetRequestedEDrx(), uCellPwrGetRequestedEDrx() and uCellPwrGetEDrx()), to determine whether deep sleep (effected by 3GPP power saving) is currently active (uCellPwrGetDeepSleepActive()) and callbacks are also available for the wake-up from deep sleep, for when the 3GPP power saving parameters change and for when the E-DRX parameters change (uCellPwrSetDeepSleepWakeUpCallback(), uCellPwrSet3gppPowerSavingCallback() and uCellPwrSetEDrxCallback()). A function is also provided to wake the module from deep sleep (uCellPwrWakeUpFromDeepSleep()) but note that this should not need to be called; waking from sleep is handled automatically by this code.

In addition, the maximum AT command "responseMaxWait" time for SARA-R5 (can be quite long when URCs get in the way of the response), which was 5 seconds, is reduced to 3 seconds, more in line with the other modules, in order to waste less time during recovery from 3GPP sleep, and the task stack for the AT client URC task is increased by 256 bytes, from 2048 to 2304.

Oh, and power saving is temporarily switched off on instance 13.0.0 of the test farm until that particular module can be replaced with one that has the production firmware version on it.