AT Client comment change only: possible consequential asserts.
Add a note that in the mutexStackPush() function there are asserts within a critical section and, with the default U_ASSERT() handler being a simple print, it is possible that if the U_ASSERT() goes off, a subsequent assert will go off in the underlying C library code because the library wants to do something like lock stdout for the print: e.g. newlib, used by GCC, would do this.