BREAKING CHANGE: New device API
Implement the first part of the new uDevice API.
This will replace all the int32_t handles we used for ble, cell, wifi and gnss
with a u_device_t handle. This will affect almost all the APIs in ubxlib.

The purpose of the new uDevice API is to isolate u-blox module initialisation with
a separate API. This will make the APIs more natural for u-blox modules that provides
more than one radio interface (such as short range modules).

Please note that this commit only adds the bare minumum for replacing the int32_t
handles with a new uDevice struct.