Cellular, Cell Locate, SARA-R5 fix: suspend UART power saving. (#482)
As noted in a previous commit, SARA-R5 can sometimes get in a bit of tizzy during Cell Locate operations. Turns out that this is due to an interaction, as yet unidentified, with UART power saving, hence with this commit, and for SARA-R5 only, UART power saving is suspended during the Cell Locate process and resumed afterwards.