New platform: SARAR5 UCPU. (#459)
This commit adds the SARA-R5 uCPU platform to ubxlib, permitting ubxlib-based applications to be written and run inside the SARA-R5 module.  This platform is NOT currently supported externally, it is purely used internally within u-blox.  It is stored here in order that it is tested and kept up to date with all the latest ubxlib developments.