Test fix for cellular only: retry C2C pairing. (#493)
On one test instance (instance 18, Zephyr + cellular SARA-R5) C2C key pairing can sometimes fail if the cellular module has recently powered on ("+CME ERROR: SEC busy" is returned by the module). In order to improve the reliability of the securityC2cBasic() test a few retries are now performed. While this could be done in the core code rather than the test code it is probably better not to perform the C2C pairing process, which involves security-related parameters, more often than is absolutely necessary, hence a decision to retry in real operation is still left up to the application.