Add support for Linux/Posix beneath Zephyr. (#488)
Add support for Linux/Posix beneath Zephyr.

Zephyr supports a "native_posix" platform, purely as a test/development host, useful because it allows tools such as address sanitizer and Valgrind to be run on the ubxlib code. This commit adds support for Zephyr on "native_posix". A few notes:

- GPIOs are not supported,
- uPortAcquireExecutableChunk() is not supported,
- on Linux/Posix Zephyr uses newlib and not its internal minimal C library,
- two UARTs are supported and will appear on a random choice of /dev/pts/x, where the value of x is printed out on the console as the first thing the application does; you may use a tool such as socat to redirect these to somewhere useful,
as part of the ubxlib test automation, a script u_run_linux.py is provided which, similar to u_run_windows.py, builds for the platform but also handles mapping of the UARTs to devices automagically based on the #defines passed into the build (for which socat must be installed).