Port event queue fixes only: errors spotted by Address Sanitizer. (#495)
There were two places in the event queue code where the uPortQueueSend() function could be passed a block of memory which was smaller than the item length defined for that queue. While this was not actively harmful (since the additional data is ignored by the function which is eventually called) it wasn't correct either and could result in a memory exception if the block of memory was at the limit of the stack or the heap. This fixes those two issues, spotted by the most excellent Address Sanitizer.