Zephyr/cell_ucpu-r5 platform change only: allow deletion of another task. (#492)
The uPortTaskDelete() call in the port API allows a task to delete itself (by passing in NULL as the parameter) or to delete another task (by passing in the handle of that task). The ubxlib code only uses the NULL form because some of the RTOSes we use (e.g. FreeRTOS in some flavours) do not support deleting anything other than the current task. However, the implementation of the uPortTaskDelete() function under Zephyr had been written to only support deletion of the current task yet this is not a requirement; Zephyr does, in fact, support deletion of another task.

Internally within u-blox we have application code that runs on the recently added cell_ucpu/r5 platform porting layer and DOES call uPortTaskDelete() with a non-NULL parameter (ThreadX, on which it normally runs, also supports deletion of another task), and we would like to be able to run this same application code on Linux-under-Zephyr, hence with this commit the restriction that only NULL is accepted as a parameter to uPortTaskDelete() for Zephyr is removed. The ability to delete another task is also now tested for Zephyr and ThreadX. FTAOD, the ubxlib code will continue to only use uPortTaskDelete(NULL).

It should be noted that using uPortTaskDelete() to delete another task requires care: the deletion process may happen at any time during execution of the task that is being deleted and so malloc()ated resources may leak, depending on the RTOS implementation mutexes could remain locked, etc.; it is up to the application design to deal with this.