Test automation change only: disable power saving on instance 18. (#496)
There is a subtle problem with the SARA-R5 module 03.15 FW and power saving mode UPSV=3, which must be used on Zephyr when the UART flow control lines are connected between the MCU and the module. This configuration is used on instance 18 on the ubxlib test farm, hence disabling cellular power saving on that instance for now.