nrf5sdk & zephyr runner: Optional startup delay for RTT
Make the startup delay for finding RTT block optional.

In the test system we now specify where the RTT block is located so there we no
longer need the delay. However, for debugging we can still need it as in this case
we use the RTT autodetection.

A small delay is also added after all the test has been printed on startup.
Without this delay we didn't get the output of the preambleHeapDefence test.