automation: Use the ELF-file to find RTT block address
Previously we have used the RTT block address autodetection mechanism provided
by Segger. However, for test HW where we flash multiple firmwares this can become
an issue since we do not power cycle the HW between sequential FW tests. In this
case there is a risk that we end up with more than one valid RTT block in the SRAM
and for this reason there is no way for the Segger autodetection mechanism to find
the right one.

To solve the problem we now parse the ELF file to find the _SEGGER_RTT symbol address.