Test automation fix for Linux/Zephyr: UART redirection
The redirection of the PTY devices created by the Linux/Zephyr build to a real TTY needs the baud rate to be specified to work correctly and also needs a udev rule to be created on the test machine to fix the connected module to a known TTY device. Note that there is a further fix, in which the devices and privileged mode are added to the docker command in Jenkinsfile: see a later commit for this.