Short range pBuf fix: memcpy() -> memmove() (#498)
uShortRangeMovePayloadFromPbufList(), part of a buffer management utility, moves data out from a set of buffers to pass it back to the caller. During this process it has to close up the gap in the buffer. For this it used memcpy(). However, where data is being copied in such a way that the destination region overlaps the source region, memmove() should be used instead, otherwise the optimization dodges used in the C library code can cause corruption of the buffer. This was happening when the test pbufInsertPayload() was run on Linux, hence memmove() is now used instead.