MQTT fix for cellular SARA-R422 and SARA-R4xxx-03B: wait for URC on disconnect. (#504)
In the MQTT client code for cellular, after AT+UMQTTC=0 has been sent to effect a disconnect from the MQTT broker, there was a wait for the +UUMQTTC URC for SARA-R5 but such a wait was deliberately omitted for a SARA-R4 series module as the AT manual has a note (section 28.6.1) that "the +UUMQTTC URC provides the result only for login and subscribe". However, this statement applies only for the SARA-R4xxx-02B series modules, the "old syntax", there should still be a wait for the +UUMQTTC URC on a disconnect for the SARA-R422 and SARA-R4xxx-03B series modules.

This was spotted because of an intermittent test failure on test instance 14 (SARA-R422) where setting keep-alive on would fail on occasion because the disconnect just before it hadn't actually completed yet 'cos the URC hadn't turned up.