Windows/U_CFG_MUTEX_DEBUG fix only: correct critical sections for U_CFG_MUTEX_DEBUG.
On Windows the critical sections are simulations of critical sections only, since critical sections as such don't exist on Windows.  The simulation is done by suspending tasks.  This causes a problem if U_CFG_MUTEX_DEBUG is defined, since when U_CFG_MUTEX_DEBUG is defined an additional mutex (gMutexList) is involved, a mutex which one of the suspended tasks may itself have locked just before it was suspended.  With this change the mutex locks protecting the critical section functions on Windows bypass the debug versions of the mutex locks (and hence don't try to lock gMutexList), allowing U_CFG_MUTEX_DEBUG to be defined for the Windows platform once more.