AT client change only: introduce uAtClientIgnoreAsync().
When an entity which uses an AT client is shutting down it may have launched asynchronous callbacks, using the facilities of the AT client, and passed those callbacks pointers to its own internal data.  Since the entity will close itself down and then close the AT client down, if any asynchronous callbacks are left in the queue to be called they may still run and access invalid pointers.  To prevent, this uAtClientIgnoreAsync() is introduced: this will cause the AT client to throw away any callbacks from that point onwards.  uAtClientIgnoreAsync() should be called before the entity that is using the AT client releases its context; it can then safely releases its context and, finally, shut down the AT client.