Cellular MQTT fix only: add delay between cellular and MQTT connection. (#509)
The internal MQTT client inside the cellular module that is used by ubxlib can take a little while to become aware that the cellular connection has been made, and it may fail to connect, returning "No Network service" (+UMQTTER: 13,34), if it is asked to connect immediately after the cellular connection is made (as indicated by UPSD). This is true both for SARA-R5, where the arrival of the +UUPSDA URC should indicate connected, and SARA-R422, where an OK response to the AT+UPSD command should indicate connected. With this commit we give it a little time to realize and 10 out of 10 runs on the automated test system showed no MQTT connect failures.