short range: Fixed EDM deadlock (#511)
This deadlock was found during the networkBle test. SPS data was received
at the same time the test was transmitting new data. When transmitting data
the first lock that will be taken is the short range mutex, after that the
EDM stream mutex. And when SPS data is received the EDM mutex will first
be taken and when the RX callback is executed the short range mutex will
be taken - obviously resulting in a deadlock.

To solve the issue the EDM mutex lock is now released during the time the
RX callback is executed.