Cellular/GNSS change: new method for indicating pin inversion. (#516)
Up to now whether a given pin (e.g. cellular PWR_ON, cellular enable power, cellular VINT, GNSS enable power, etc.) is inverted or not has been controlled via a compile time option (e.g. U_CELL_PWR_ON_PIN_INVERTED). With this commit it is possible to indicate that the sense of a pin is inverted in the calls to uCellAdd(), uCellPwrResetHard(), uCellPwrSetDtrPowerSavingPin() and uGnssAdd() by ORing the pin with U_CELL_PIN_INVERTED (for the cellular cases) or U_GNSS_PIN_INVERTED (for the GNSS cases).

This is now the preferred mechanism for indicating pin inversion, since then ubxlib can be compiled as a static library that can be told what to by the application through an API.

Note that you should NOT use both the compile-time and the "ORing with U_xxxx_PIN_INVERTED" mechanisms at the same time as that will result in the sense of the pin being inverted twice, which is not what you want.