MQTT change only: setting/getting local port. (#518)
The localPort entry in uMqttClientConnection_t and the uCellMqttSetLocalPort()/uCellMqttGetLocalPort() APIs were implemented such that, if the local port is not known to ubxlib, 1883 would be returned for a non-secure connection and 8883 for a secure connection. This is, of course, complete rubbish; for cases where setting of the local port is not possible (i.e. for any module type other than cellular SARA-R412M-02B) the local port is whatever the TCP implementation in the module has chosen; it is not fixed to 1883/8883, those ports are the ports of the far-end only and the local port simply cannot be read from the module.

With this commit U_ERROR_COMMON_NOT_SUPPORTED is returned by uCellMqttGetLocalPort() if either (a) setting of the local port is not supported or (b) in the case of SARA-R412M-02B if no port has explicitly been set with uCellMqttSetLocalPort() (since SARA-R412M-02B does not return a port number unless one has explicitly been set).