BREAKING CHANGE IF YOU HAVE YOUR OWN PORT: add logging on/off functions to porting layer. (#515)
With this commit a new pair of functions are added to the port API: uPortLogOn()/uPortLogOff(). With the addition of these functions, as well as there being compile-time control of logging through the U_CFG_ENABLE_LOGGING macro, logging can now also be controlled by the application, allowing the ubxlib code to be built as a static library.

It is NOT a requirement that these two functions are implemented: where they are not implemented U_ERROR_COMMON_NOT_IMPLEMENTED should be returned.

Note that finer control of debug logging still remains possible in the usual way, i.e. the AT client, source of most debug prints, continues to permit printing of AT commands/responses and AT client debug traces to be controlled via the uAtClientDebugSet() and uAtClientPrintAtSet() functions; the uPortLogOn()/uPortLogOff() functions simply provide a lower-level override for all logging out of ubxlib.