automation: Add analyze tasks for stm32cube and nrfconnect
Added an "analyze" task for stm32cube and nrfconnect that will run Ericsson
CodeChecker on ubxlib source. CodeChecker uses clang-tidy (linter) and
clang analyze (static code analyzer).

The output will be placed in the <build_dir>/analyze_html.

No analyze task was added for ESP or nRF5 SDK targets due to these reasons:

ESP:
Clang currently doesn't support Xtensa (clang-14). Espressif have made
an implementation for it but it hasn't been included in the official
clang repo yet.

nRF5 SDK:
Clang had a lot of issues with the Makefile that is included from nRF5 SDK.
Maybe it is possible to solve, but I didn't think it was worth the effort.