debug_utils: Add dump thread support for ESP32
Please note that current thread will not be dumped correctly.
But this shouldn't be an issue when used with the inactivity
detector, since in this case the current thread will be the
inactivity task.

If there are a need to get a correct backtrace for current thread
this can be done by using esp_backtrace_get_start(). But in this
case we should also fix the ARM backtrace.