New uDevice APIs implemented
uDeviceOpen and uDeviceClose are now implemented.
As a first step these just sit on top of the old APIs in order to temporarily maintain backwards compatibility  and facilitate the transformation of all test programs. Once this transformation has been made the actual implementation will change and the old APIs will be removed.
Also for compatibility reasons new APIs named uNetworkInterfaceUp and uNetworkInterfaceDown have been added. These are intended to replace uNetworkUp and uNetworkDown. The main difference is that these specify which network interface to operate. This is crucial for module with multiple interfaces (short-range)
Another new API is uNetworkConfigure which is a replacement for specifying the network specific options of the old uNetworkAdd