Cellular change for SARA-R5 only: fix authentication mode command (#536)
The authentication mode command (AT+UAUTHREQ), used when a username and password are required to activate a PDP context, was originally written for SARA-R4 where the username and password parameters are swapped versus the other u-blox module types. This was never corrected when SARA-R5 was added; it is now fixed.