vscode: Use Python script to get ARM toolchain path
There was sometimes an issue with the "Command Variable" extension that
it returned "Unknown" when getting the ARM toolchain from arm-toolchain-path.txt.
For this reason we will now test a new approach. By using the "Tasks Shell Input"
extension instead it is possible to read a value from a shell command and use
it as an variable in launch.json and tasks.json. A small Python script has been
added that will return the ARM toolchain path via u_packages.