Cellular change only: re-establish profile on return to coverage.
When a cellular module falls out of coverage it will always, automatically, attempt to regain coverage.  However, the loss of coverage resets the internal "profile", basically the PDP context but internal to the module, and this profile is NOT re-established on return to service in all cases.  With this commit the ubxlib code will re-establish the profile when service is regained so that any IP connection or MQTT broker connection may be re-established.

IMPORTANT: an IP connection or MQTT broker connection made by the application is NOT, cannot be, re-established automatically by the ubxlib code; the ubxlib code does not have the necessary information to do so.  Hence it is up to the application to re-establish such connections, e.g. by setting a callback with uCellNetSetRegistrationStatusCallback() and then re-establishing connections in cases where service has been lost and then regained.