Test automation change: add ubxlib.h checker. (#551)
A script is added, u_run_check_ubxlib_h.py, run as instance 8 in DATABASE.md, which checks that any header file found in an API directory of ubxlib is either included in ubxlib.h or is instead added to a "known excludes" list, ubxlib_h_excludes.txt.  This is to make it less likely (though it is not a fool-proof check) that we miss one out, which would be embarrassing.