BEHAVIOUR CHANGE FOR MQTT/MQTT-SN, cellular only: retry on failures due to radio conditions. (#549)
MQTT-SN uses a UDP bearer; this is less reliable than the TCP bearer used by MQTT hence it is worth retrying on a failure due to radio conditions. There is no harm in doing this for MQTT also, hence with this commit we retry the radio-based MQTT operations on failure ONLY IF the error cause given by the cellular module is due to radio conditions.

To return to the old behaviour, with no retries, uCellMqttSetRetries() can be called with 0 for numRetries.