Fix to test code only: less reliance on timing in test portOs. (#561)
In the test portOs a task is started and then does some trylocking/unlocking/locking of a mutex; the task which created it also locking/unlocking the same mutex, to check that the mutex locking works. However, the sequence relies on timing, immediacy, and Windows doesn't really do that, so with this commit we remove the bit of the test that relies on timing if we're running on Windows. There is still testing that mutexes work, just a little simpler for the Windows case.