Fix to test code in mqttClient(): messageIndicationCallback() should get pointer to static. (#565)
In the test mqttClient() a message indication callback is set up to be passed a pointer to a stack variable. This is fine except that there seems to be a bug in SARA-R422 where the module sometimes doesn't send the URC indicating a successful MQTT publish while it is in 32 kHz sleep. We wait 30 seconds for the publish to succeed and, when we hear nothing, we send an AT command to get the MQTT error code. This wakes the module up and the URC arrives, so now we've (a) failed the publish and (b) the message indication callback has been called. This is also fine except that, because the test has failed at the publish step, the mqttClient() function exits at the U_PORT_TEST_ASSERT() of the publish return value and we move on to the next test but messageIndicationCallback() is still running, now with a parameter pointing at a bit of unwound stack.

That's not so fine.

This commit changes the test so that the parameter passed to messageIndicationCallback() is a pointer to a static variable rather than a stack variable, making it safe.