Cell Locate change to test code only: reboot module if no PWR_ON pin. (#568)
The cellular test postamble powers the module down but only if there is a PWR_ON pin (otherwise we can't power the module back up again). In the particular case of the cell locate test, we occasionally see SARA-R5 become unresponsive to AT commands in the following test (at the point where the power-on sequence issues AT+CFUN=4 after configuring the module), and of course there is no way to get the module out of that state without the PWR_ON pin to power it off.

Hence with this commit, if there is no PWR_ON pin, then, after we complete the cell locate test, we perform a reboot of the module in place of what would normally be the power off.