BEHAVIOUR CHANGE IN AT CLIENT FUNCTION uAtClientWaitCharacter(), fix for cell [sockets]. (#570)
While the [rarely used] AT client function uAtClientWaitCharacter() is waiting for a single character to arrive we could miss URCs that just happened to land at the same time. This is because the function did not wait for URCs correctly, more usually consuming any intervening characters up to the wanted character. This is a problem because a URC can occur at any time; in particular, where the uAtClientWaitCharacter() is used to wait for the @ character during cellular sockets operations, it is possible for a +UUSORD URC (indicating that downlink data is available) to be emitted immediately before the @ character. When such a URC is missed we will not notice that downlink data is available, which is not good. We have seen this, very occasionally, in the sockets loopback tests on cellular.

This commit modifies uAtClientWaitCharacter() so that it does NOT throw intervening characters away where they could form part of a URC that we are meant to be looking for, correctly processing them and calling URC handlers as appropriate.

Note: as a corollary of this change uAtClientWaitCharacter() no longer accepts 0x0d or 0x0a as wait-for-characters since 0x0d/0x0a will always precede a URC and so form part of the URC handling process.