Device API change: add powerOff parameter to uDeviceClose().
The new uDeviceClose() function, for cellular and GNSS, always performs a power-down, while for ShortRange modules powering down is not actually supported, only a logical disconnection occurs.  The equivalent old uNetworkRemove() API never performed a power-down (that was done on uNetworkDown() for the cellular and GNSS cases).  With the new behaviour the test time has increased from 43 minutes to 53 minutes, since powering down can be a time-consuming process and it is done many 10s of times.

With this commit a parameter powerOff is added to uDeviceClose().  This allows the user to get the old behaviour, and should also return the test time back to what it was previously.  If the parameter is set to true for a ShortRange device an error is returned.