Arduino platform change only: add missing u_cell_gpio.c. (#578)
The file that implements the cellular GPIO functions, i.e. allowing GPIOs that are over on the cellular module to be controlled from ubxlib, was missing from the source.txt file that is used by the Python script that pulls together all of the source files for Arduino.  This would normally cause a compilation error except that the file that tests it was also missed out of source_test.txt.  D'oh.  Both are now added.