AT client change only: add hysteresis. (#577)
The uAtClientWaitCharacter() fix in commit b0d02f77bb64cf4704fef999a93d3559d842afda is designed to make sure that URCs are not lost when waiting for a single character. However, it relies on at least the prefix of the URC arriving in one go. I had thought there was hysteresis in the function of the AT client (bufferFill()) that reads from the UART in order to make this happen but I was mistaken; it was the design intention but it did not actually work. This commit fixes that.