Fix in Wi-Fi part of network and change of uDeviceHandle_t definition.
The new function uNetworkPrivateChangeStateWifi(), at the end of the bring-up part of the process, calls uWifiSetNetworkStatusCallback().  The first parameter to that function should be of type uDeviceHandle_t (devHandle) but queueHandle (type uPortQueueHandle_t) was being passed instead; many of the static functions in that file do use queueHandle so it is an easy mistake to make.  The compiler didn't throw an error or even a warning on this since the underlying type of the parameter is a void *.  CodeChecker didn't complain about this either and, it would seem from this stack overflow discussion:

https://stackoverflow.com/questions/72415062/c-compiler-checking-of-a-typedefed-void

...there is no way to make a C compiler complain about this; typedef'ing a void * is, actually, completely pointless, maybe even actively dangerous.

As some measure of protection the definition of uDeviceHandle_t is now changed to be a pointer to the actual uDeviceCfg_t structure and the user is just told not to go de-referencing it.